fix(jsonschema): preserve 64-bit integer tag constants on 32-bit
parseNumericValue and parseTypedScalar returned int(n) for signed
integer kinds, truncating int64 const/enum/examples/len values above
2^31-1 on a 32-bit build. Return int64 instead; the numeric comparison
handles every integer kind via reflection, so generated JSON and
validation are unchanged on 64-bit. Update the assertions that pinned
the Go type to int.

+++ jsonschema/generate_test.go
@@ -2107,7 +2107,7 @@ func TestGenerateFor_JSONSchemaTag_Examples(t *testing.T) {
 	require.NoError(t, err)
 
 	// Integer examples should be parsed as integers (precision-preserving).
-	assert.Equal(t, []any{8080, 3000, 443}, s.Properties["port"].Examples)
+	assert.Equal(t, []any{int64(8080), int64(3000), int64(443)}, s.Properties["port"].Examples)
 	// String examples should be strings.
 	assert.Equal(t, []any{"debug", "release"}, s.Properties["mode"].Examples)
 }
